# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateAnnouncementRequest": {
        "additionalProperties": false,
        "properties": {
          "audience_condition": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Audience Condition",
            "type": "object"
          },
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/AnnouncementCategory"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "title": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "body",
          "category"
        ],
        "title": "CreateAnnouncementRequest",
        "type": "object"
      },
...
      "EditAnnouncementRequest": {
        "additionalProperties": false,
        "properties": {
          "audience_condition": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Audience Condition",
            "type": "object"
          },
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/AnnouncementCategory"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "title": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "body",
          "category",
          "expected_version"
        ],
        "title": "EditAnnouncementRequest",
        "type": "object"
      },
...
      "Envelope_AnnouncementListResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnnouncementListResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "Envelope[AnnouncementListResponse]",
        "type": "object"
      },
      "Envelope_AnnouncementResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnnouncementResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "Envelope[AnnouncementResponse]",
        "type": "object"
      },
      "Envelope_AudienceSnapshotResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AudienceSnapshotResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "request_id"
        ],
        "title": "Envelope[AudienceSnapshotResponse]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublishAnnouncementRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmation_token": {
            "minLength": 1,
            "title": "Confirmation Token",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version",
          "confirmation_token"
        ],
        "title": "PublishAnnouncementRequest",
        "type": "object"
      },
      "RejectAnnouncementRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "RejectAnnouncementRequest",
        "type": "object"
      },
...
      "ScheduleAnnouncementRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmation_token": {
            "minLength": 1,
            "title": "Confirmation Token",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "scheduled_at": {
            "format": "date-time",
            "title": "Scheduled At",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "confirmation_token",
          "scheduled_at"
        ],
        "title": "ScheduleAnnouncementRequest",
        "type": "object"
      },
...
      "VersionedActionRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "VersionedActionRequest",
        "type": "object"
      },
...
      "WithdrawAnnouncementRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "WithdrawAnnouncementRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "物业社区管理智能体 — 统一后端服务",
    "title": "Property Community Management Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/announcements": {
      "get": {
        "operationId": "search_api_announcements_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "announcement"
        ]
      },
      "post": {
        "operationId": "create_api_announcements_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnouncementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}": {
      "get": {
        "operationId": "get_api_announcements__announcement_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get",
        "tags": [
          "announcement"
        ]
      },
      "patch": {
        "operationId": "edit_api_announcements__announcement_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditAnnouncementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}/actions/approve": {
      "post": {
        "operationId": "approve_api_announcements__announcement_id__actions_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionedActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}/actions/publish": {
      "post": {
        "operationId": "publish_api_announcements__announcement_id__actions_publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishAnnouncementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}/actions/reject": {
      "post": {
        "operationId": "reject_api_announcements__announcement_id__actions_reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectAnnouncementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}/actions/schedule": {
      "post": {
        "operationId": "schedule_publish_api_announcements__announcement_id__actions_schedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleAnnouncementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedule Publish",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}/actions/withdraw": {
      "post": {
        "operationId": "withdraw_api_announcements__announcement_id__actions_withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawAnnouncementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}/audience-preview": {
      "get": {
        "operationId": "preview_audience_api_announcements__announcement_id__audience_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AudienceSnapshotResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Audience",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}/submit-review": {
      "post": {
        "operationId": "submit_review_api_announcements__announcement_id__submit_review_post",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionedActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnnouncementResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Review",
        "tags": [
          "announcement"
        ]
      }
    },
    "/api/announcements/{announcement_id}/versions": {
      "get": {
        "operationId": "versions_api_announcements__announcement_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "announcement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Announcement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_AnnouncementVersionResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Versions",
        "tags": [
          "announcement"
        ]
      }
    },
...
  }
}